# Doubled rehash performance